Fix automatic JavaScript placeholder compatibility
