--- .github/workflows/fix-js-placeholder-compat.yml
@@ -1,23 +1,0 @@
-name: Apply JavaScript placeholder compatibility fixes
-on:
-  push:
-    branches: [security/javascript-placeholder-bindings]
-permissions:
-  contents: write
-jobs:
-  patch:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: security/javascript-placeholder-bindings
-      - name: Apply fixes
-        run: |
-          python3 .github/fix_js_placeholder_compat.py
-          rm -f .github/fix_js_placeholder_compat.py .github/workflows/fix-js-placeholder-compat.yml
-          git config user.name 'Ben'
-          git config user.email 'benbergen12@gmail.com'
-          git add -A
-          git commit -m 'Fix automatic JavaScript placeholder compatibility'
-          git push origin HEAD:security/javascript-placeholder-bindings
